pyproject.toml
[project]
name = "visa-bot"
version = "0.1.0"
description = "AI-powered CLI visa application assistant"
requires-python = ">=3.11"
dependencies = [
    "groq>=0.9.0",
    "rich>=13.7.0",
    "requests>=2.31.0",
    "trafilatura>=1.8.0",
    "duckduckgo-search>=6.0.0",
    "chromadb>=0.5.0",
    "python-dotenv>=1.0.0",
]

[project.scripts]
visa-bot = "main:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["core", "knowledge", "llm", "retrieval", "ui"]
